[Chore] #186 - 파일 분리
