+++ WSSiOS/WSSiOS/Source/Presentation/MyPage/MyPageView/MyPageDeleteIDView.swift
@@ -12,26 +12,13 @@ import Then
 
 final class MyPageDeleteIDView: UIView {
     
-    //MARK: - Properties
-    
-    
     //MARK: - Components
     
     private let scrollView = UIScrollView()
     private let contentView = UIView()
     
-    private let reasonView = UIView()
-    private let reasonTitleLabel = UILabel()
-    var reasonTableView = UITableView(frame: .zero, style: .plain)
-    var reasonTextView = UITextView()
-    private var reasonTextViewPlaceholder = UILabel()
-    var reasonCountLabel = UILabel()
-    private let reasonCountLimitLabel = UILabel()
-    
-    private let checkView = UIView()
-    private let checkTitleLabel = UILabel()
-    var checkTableView = UITableView(frame: .zero, style: .plain)
-    
+    let reasonView = MyPageDeleteIDReasonView()
+    let checkView = MyPageDeleteIDCheckView()
     lazy var agreeDeleteIDButton = UIButton()
     private let agreeDeleteIDLabel = UILabel()
     
@@ -61,59 +48,6 @@ final class MyPageDeleteIDView: UIView {
         
         scrollView.do {
             $0.showsVerticalScrollIndicator = false
-        }
-        
-        reasonTitleLabel.do {
-            $0.font = .HeadLine1
-            $0.textColor = .wssBlack
-            $0.makeAttribute(with: StringLiterals.MyPage.DeleteID.reasonTitle)?
-                .lineHeight(1.4)
-                .kerning(kerningPixel: -1.2)
-                .partialColor(color: .wssPrimary100, rangeString: StringLiterals.MyPage.DeleteID.reasonTitleColor)
-                .applyAttribute()
-        }
-        
-        reasonTableView.do {
-            $0.isScrollEnabled = false
-            $0.showsVerticalScrollIndicator = false
-            $0.separatorStyle = .none
-        }
-        
-        reasonTextView.do {
-            $0.font = .Body2
-            $0.textColor = .wssBlack
-            $0.backgroundColor = .wssGray50
-            $0.layer.cornerRadius = 14
-            $0.textContainerInset = UIEdgeInsets(top: 10, left: 16, bottom: 12, right: 16)
-            $0.spellCheckingType = .no
-            $0.autocorrectionType = .no
-            $0.autocapitalizationType = .none
-        }
-        
-        reasonTextViewPlaceholder.do {
-            $0.applyWSSFont(.body2, with: StringLiterals.MyPage.DeleteID.reasonPlaceHolder)
-            $0.textColor = .wssGray200
-        }
-        
-        reasonCountLabel.do {
-            $0.applyWSSFont(.label1, with: "0")
-            $0.textColor = .wssGray300
-        }
-        
-        reasonCountLimitLabel.do {
-            $0.applyWSSFont(.label1, with: "/" + String(MyPageDeleteIDViewModel.textViewMaxLimit))
-            $0.textColor = .wssGray200
-        }
-        
-        checkTitleLabel.do {
-            $0.applyWSSFont(.headline1, with: StringLiterals.MyPage.DeleteID.checkTitle)
-            $0.textColor = .wssBlack
-        }
-        
-        checkTableView.do {
-            $0.isScrollEnabled = false
-            $0.showsVerticalScrollIndicator = false
-            $0.separatorStyle = .none
         }
         
         agreeDeleteIDButton.do {
@@ -147,14 +81,6 @@ final class MyPageDeleteIDView: UIView {
                                 agreeDeleteIDButton,
                                 agreeDeleteIDLabel,
                                 completeButton)
-        reasonView.addSubviews(reasonTitleLabel,
-                               reasonTableView,
-                               reasonTextView,
-                               reasonCountLimitLabel,
-                               reasonCountLabel)
-        reasonTextView.addSubview(reasonTextViewPlaceholder)
-        checkView.addSubviews(checkTitleLabel,
-                              checkTableView)
     }
     
     private func setLayout() {
@@ -174,54 +100,11 @@ final class MyPageDeleteIDView: UIView {
         reasonView.snp.makeConstraints {
             $0.top.equalToSuperview().inset(45)
             $0.leading.trailing.equalToSuperview().inset(20)
-            
-            reasonTitleLabel.snp.makeConstraints {
-                $0.top.leading.equalToSuperview()
-            }
-            
-            reasonTableView.snp.makeConstraints {
-                $0.top.equalTo(reasonTitleLabel.snp.bottom).offset(20)
-                $0.leading.trailing.equalToSuperview()
-                $0.height.equalTo(44 * 5)
-            }
-            
-            reasonTextView.snp.makeConstraints {
-                $0.top.equalTo(reasonTableView.snp.bottom).offset(2)
-                $0.width.equalToSuperview()
-                $0.height.equalTo(114)
-                
-                reasonTextViewPlaceholder.snp.makeConstraints {
-                    $0.top.equalToSuperview().inset(10)
-                    $0.leading.trailing.equalToSuperview().inset(16)
-                }
-            }
-            
-            reasonCountLimitLabel.snp.makeConstraints {
-                $0.top.equalTo(reasonTextView.snp.bottom).offset(4)
-                $0.trailing.equalToSuperview()
-            }
-            
-            reasonCountLabel.snp.makeConstraints {
-                $0.top.equalTo(reasonCountLimitLabel.snp.top)
-                $0.trailing.equalTo(reasonCountLimitLabel.snp.leading)
-                $0.bottom.equalToSuperview()
-            }
         }
         
         checkView.snp.makeConstraints {
             $0.top.equalTo(reasonView.snp.bottom).offset(60)
             $0.leading.trailing.equalToSuperview().inset(20)
-            
-            checkTitleLabel.snp.makeConstraints {
-                $0.top.leading.equalToSuperview()
-            }
-            
-            checkTableView.snp.makeConstraints {
-                $0.top.equalTo(checkTitleLabel.snp.bottom).offset(20)
-                $0.leading.trailing.equalToSuperview()
-                $0.height.equalTo(318)
-                $0.bottom.equalToSuperview()
-            }
         }
         
         agreeDeleteIDButton.snp.makeConstraints {
@@ -245,20 +128,8 @@ final class MyPageDeleteIDView: UIView {
 }
 
 extension MyPageDeleteIDView {
-
-    //MARK: - Data
     
-    func bindTextCount(count: Int) {
-        reasonCountLabel.do {
-            $0.applyWSSFont(.label1, with: String(count))
-        }
-    }
-    
-    func bindText(text: String) {
-        reasonTextView.do {
-            $0.text = text
-        }
-    }
+    //MARK: - Custom Method
     
     func agreeDeleteIDButtonIsSelected(isSelected: Bool) {
         agreeDeleteIDButton.setImage( isSelected ? .checkSelected : .checkDefault , for: .normal)
@@ -267,10 +138,6 @@ extension MyPageDeleteIDView {
     func completeButtonIsEnabled(isEnabled: Bool) {
         completeButton.backgroundColor = isEnabled ? .Primary100 : .wssGray70
         completeButton.isEnabled = isEnabled
-    }
-    
-    func placeholderIsHidden(isHidden: Bool) {
-        reasonTextViewPlaceholder.isHidden = isHidden
     }
 }
